// FrontBookRentals/src/app/api-client/model/clientResponseDtoGenericResponse.ts
/**
 * Book Rentals API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ClientResponseDto } from './clientResponseDto';


export interface ClientResponseDtoGenericResponse { 
    success: boolean;
    message: string;
    data?: ClientResponseDto;
}