// FrontBookRentals/src/app/api-client/model/guidGenericResponse.ts
/**
 * Book Rentals API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface GuidGenericResponse { 
    success: boolean;
    message: string;
    data?: string;
}